refactor(ribir): 💡 add a mock backend
