+++ ribir/src/lib.rs
@@ -2,9 +2,9 @@ pub use ribir_core as core;
 #[cfg(feature = "widgets")]
 pub use ribir_widgets as widgets;
 pub mod app;
+mod backends;
 pub mod timer;
 mod winit_shell_wnd;
-
 #[cfg(feature = "material")]
 pub use ribir_material as material;
